feat(importer): cleanup
